+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -238,6 +238,7 @@ go_library(
         "//cmd/frontend/hubspot",
         "//cmd/frontend/hubspot/hubspotutil",
         "//cmd/frontend/internal/cloneurls",
+        "//cmd/frontend/internal/cody",
         "//cmd/frontend/internal/inventory",
         "//cmd/frontend/internal/processrestart",
         "//cmd/frontend/internal/search/logs",
@@ -254,7 +255,6 @@ go_library(
         "//internal/codeintel/dependencies",
         "//internal/codeintel/dependencies/shared",
         "//internal/codeintel/resolvers",
-        "//internal/cody",
         "//internal/codygateway",
         "//internal/conf",
         "//internal/conf/conftypes",
