+++ cmd/frontend/internal/app/jscontext/BUILD.bazel
@@ -13,11 +13,11 @@ go_library(
         "//cmd/frontend/graphqlbackend",
         "//cmd/frontend/internal/app/assetsutil",
         "//cmd/frontend/internal/app/ui/sveltekit",
+        "//cmd/frontend/internal/cody",
         "//cmd/frontend/webhooks",
         "//internal/actor",
         "//internal/auth/providers",
         "//internal/auth/userpasswd",
-        "//internal/cody",
         "//internal/conf",
         "//internal/conf/deploy",
         "//internal/database",
