+++ cmd/frontend/internal/cody/BUILD.bazel
@@ -1,5 +1,5 @@
-load("//dev:go_defs.bzl", "go_test")
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("//dev:go_defs.bzl", "go_test")
 
 go_library(
     name = "cody",
@@ -9,8 +9,8 @@ go_library(
         "subscription.go",
         "utils.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/cody",
-    visibility = ["//:__subpackages__"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/cody",
+    visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//cmd/frontend/backend",
         "//internal/accesstoken",
