+++ cmd/frontend/internal/dotcom/productsubscription/BUILD.bazel
@@ -26,12 +26,12 @@ go_library(
     deps = [
         "//cmd/frontend/graphqlbackend",
         "//cmd/frontend/graphqlbackend/graphqlutil",
+        "//cmd/frontend/internal/cody",
         "//internal/accesstoken",
         "//internal/actor",
         "//internal/audit",
         "//internal/auth",
         "//internal/authz",
-        "//internal/cody",
         "//internal/codygateway",
         "//internal/completions/client/anthropic",
         "//internal/completions/client/fireworks",
@@ -93,11 +93,11 @@ go_test(
     ],
     deps = [
         "//cmd/frontend/graphqlbackend",
+        "//cmd/frontend/internal/cody",
         "//internal/accesstoken",
         "//internal/actor",
         "//internal/audit/audittest",
         "//internal/authz",
-        "//internal/cody",
         "//internal/conf",
         "//internal/database",
         "//internal/database/dbmocks",
