+++ cmd/frontend/internal/dotcom/productsubscription/codygateway_dotcom_user.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"math"
 
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/cody"
 	"github.com/sourcegraph/sourcegraph/internal/auth"
-	"github.com/sourcegraph/sourcegraph/internal/cody"
 	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/featureflag"
