+++ cmd/frontend/internal/httpapi/BUILD.bazel
@@ -26,6 +26,7 @@ go_library(
         "//cmd/frontend/backend",
         "//cmd/frontend/enterprise",
         "//cmd/frontend/graphqlbackend",
+        "//cmd/frontend/internal/cody",
         "//cmd/frontend/internal/handlerutil",
         "//cmd/frontend/internal/httpapi/releasecache",
         "//cmd/frontend/internal/httpapi/webhookhandlers",
@@ -40,7 +41,6 @@ go_library(
         "//internal/auth",
         "//internal/authz",
         "//internal/codeintel/types",
-        "//internal/cody",
         "//internal/conf",
         "//internal/cookie",
         "//internal/database",
