+++ cmd/frontend/internal/httpapi/completions/chat.go
@@ -5,8 +5,8 @@ import (
 
 	"net/http"
 
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/cody"
 	sgactor "github.com/sourcegraph/sourcegraph/internal/actor"
-	"github.com/sourcegraph/sourcegraph/internal/cody"
 	"github.com/sourcegraph/sourcegraph/internal/dotcom"
 
 	"github.com/sourcegraph/log"
